requirements.txt
streamlit
deepface
opencv-python
pillow
tensorflow==2.12.0
